Move metadata compaction code into submodules
