+++ crates/contracts/ab-contracts-io-type/src/metadata.rs
@@ -1,3 +1,6 @@
+mod compact;
+
+use crate::metadata::compact::compact_metadata;
 use core::ptr;
 
 /// Max capacity for metadata bytes used in fixed size buffers
@@ -40,8 +43,8 @@ pub const fn concat_metadata_sources(sources: &[&[u8]]) -> ([u8; MAX_METADATA_CA
 /// Metadata encoding consists of this enum variant treated as `u8` followed by optional metadata
 /// encoding rules specific to metadata type variant (see variant's description).
 ///
-/// This metadata is sufficient to fully reconstruct hierarchy of the type in order to generate
-/// language bindings, auto-generate UI forms, etc.
+/// This metadata is enough to fully reconstruct the hierarchy of the type to generate language
+/// bindings, auto-generate UI forms, etc.
 ///
 /// [`TrivialType::METADATA`]: crate::trivial_type::TrivialType::METADATA
 /// [`IoType::METADATA`]: crate::IoType::METADATA
@@ -519,18 +522,20 @@ impl IoTypeMetadataKind {
         })
     }
 
+    // TODO: Create wrapper type for metadata bytes and move this method there
     /// Produce compact metadata.
     ///
-    /// Compact metadata retains the shape, but throws some of the details. Specifically following
+    /// Compact metadata retains the shape, but throws some details. Specifically, the following
     /// transformations are applied to metadata:
-    /// * Struct names, enum names and enum variant names are removed (replaced with 0 bytes names)
+    /// * Struct names, enum names and enum variant names are removed (replaced with zero bytes
+    ///   names)
     /// * Structs and enum variants are turned into tuple variants (removing field names)
     ///
     /// This is typically called by higher-level functions and doesn't need to be used directly.
     ///
     /// This function takes an `input` that starts with metadata defined in [`IoTypeMetadataKind`]
     /// and `output` where compact metadata must be written. Since input might have other data past
-    /// the data structure to be processed, remainders of input and output are returned back to the
+    /// the data structure to be processed, the remainder of input and output are returned to the
     /// caller.
     ///
     /// Unexpected metadata kind results in `None` being returned.
@@ -540,549 +545,4 @@ impl IoTypeMetadataKind {
     ) -> Option<(&'i [u8], &'o mut [u8])> {
         compact_metadata(input, output)
     }
-}
-
-/// This macro is necessary to reduce boilerplate due to lack of `?` in const environment
-macro_rules! forward_option {
-    ($expr:expr) => {{
-        let Some(result) = $expr else {
-            return None;
-        };
-        result
-    }};
-}
-
-const fn compact_metadata<'i, 'o>(
-    mut input: &'i [u8],
-    mut output: &'o mut [u8],
-) -> Option<(&'i [u8], &'o mut [u8])> {
-    if input.is_empty() || output.is_empty() {
-        return None;
-    }
-
-    let kind = forward_option!(IoTypeMetadataKind::try_from_u8(input[0]));
-
-    match kind {
-        IoTypeMetadataKind::Unit
-        | IoTypeMetadataKind::Bool
-        | IoTypeMetadataKind::U8
-        | IoTypeMetadataKind::U16
-        | IoTypeMetadataKind::U32
-        | IoTypeMetadataKind::U64
-        | IoTypeMetadataKind::U128
-        | IoTypeMetadataKind::I8
-        | IoTypeMetadataKind::I16
-        | IoTypeMetadataKind::I32
-        | IoTypeMetadataKind::I64
-        | IoTypeMetadataKind::I128 => copy_n_bytes(input, output, 1),
-        IoTypeMetadataKind::Struct => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, None, false)
-        }
-        IoTypeMetadataKind::Struct0 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(0), false)
-        }
-        IoTypeMetadataKind::Struct1 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct1 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(1), false)
-        }
-        IoTypeMetadataKind::Struct2 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct2 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(2), false)
-        }
-        IoTypeMetadataKind::Struct3 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct3 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(3), false)
-        }
-        IoTypeMetadataKind::Struct4 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct4 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(4), false)
-        }
-        IoTypeMetadataKind::Struct5 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct5 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(5), false)
-        }
-        IoTypeMetadataKind::Struct6 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct6 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(6), false)
-        }
-        IoTypeMetadataKind::Struct7 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct7 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(7), false)
-        }
-        IoTypeMetadataKind::Struct8 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct8 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(8), false)
-        }
-        IoTypeMetadataKind::Struct9 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct9 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(9), false)
-        }
-        IoTypeMetadataKind::Struct10 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct10 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(10), false)
-        }
-        IoTypeMetadataKind::Struct11 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct11 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(11), false)
-        }
-        IoTypeMetadataKind::Struct12 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct12 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(12), false)
-        }
-        IoTypeMetadataKind::Struct13 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct13 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(13), false)
-        }
-        IoTypeMetadataKind::Struct14 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct14 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(14), false)
-        }
-        IoTypeMetadataKind::Struct15 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct15 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(15), false)
-        }
-        IoTypeMetadataKind::Struct16 => {
-            // Convert struct with field names to tuple struct
-            output[0] = IoTypeMetadataKind::TupleStruct16 as u8;
-            (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-            compact_struct(input, output, Some(16), false)
-        }
-        IoTypeMetadataKind::TupleStruct => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, None, true)
-        }
-        IoTypeMetadataKind::TupleStruct1 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(1), true)
-        }
-        IoTypeMetadataKind::TupleStruct2 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(2), true)
-        }
-        IoTypeMetadataKind::TupleStruct3 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(3), true)
-        }
-        IoTypeMetadataKind::TupleStruct4 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(4), true)
-        }
-        IoTypeMetadataKind::TupleStruct5 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(5), true)
-        }
-        IoTypeMetadataKind::TupleStruct6 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(6), true)
-        }
-        IoTypeMetadataKind::TupleStruct7 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(7), true)
-        }
-        IoTypeMetadataKind::TupleStruct8 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(8), true)
-        }
-        IoTypeMetadataKind::TupleStruct9 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(9), true)
-        }
-        IoTypeMetadataKind::TupleStruct10 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(10), true)
-        }
-        IoTypeMetadataKind::TupleStruct11 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(11), true)
-        }
-        IoTypeMetadataKind::TupleStruct12 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(12), true)
-        }
-        IoTypeMetadataKind::TupleStruct13 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(13), true)
-        }
-        IoTypeMetadataKind::TupleStruct14 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(14), true)
-        }
-        IoTypeMetadataKind::TupleStruct15 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(15), true)
-        }
-        IoTypeMetadataKind::TupleStruct16 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_struct(input, output, Some(16), true)
-        }
-        IoTypeMetadataKind::Enum => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, None, true)
-        }
-        IoTypeMetadataKind::Enum1 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(1), true)
-        }
-        IoTypeMetadataKind::Enum2 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(2), true)
-        }
-        IoTypeMetadataKind::Enum3 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(3), true)
-        }
-        IoTypeMetadataKind::Enum4 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(4), true)
-        }
-        IoTypeMetadataKind::Enum5 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(5), true)
-        }
-        IoTypeMetadataKind::Enum6 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(6), true)
-        }
-        IoTypeMetadataKind::Enum7 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(7), true)
-        }
-        IoTypeMetadataKind::Enum8 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(8), true)
-        }
-        IoTypeMetadataKind::Enum9 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(9), true)
-        }
-        IoTypeMetadataKind::Enum10 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(10), true)
-        }
-        IoTypeMetadataKind::Enum11 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(11), true)
-        }
-        IoTypeMetadataKind::Enum12 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(12), true)
-        }
-        IoTypeMetadataKind::Enum13 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(13), true)
-        }
-        IoTypeMetadataKind::Enum14 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(14), true)
-        }
-        IoTypeMetadataKind::Enum15 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(15), true)
-        }
-        IoTypeMetadataKind::Enum16 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(16), true)
-        }
-        IoTypeMetadataKind::EnumNoFields => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, None, false)
-        }
-        IoTypeMetadataKind::EnumNoFields1 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(1), false)
-        }
-        IoTypeMetadataKind::EnumNoFields2 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(2), false)
-        }
-        IoTypeMetadataKind::EnumNoFields3 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(3), false)
-        }
-        IoTypeMetadataKind::EnumNoFields4 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(4), false)
-        }
-        IoTypeMetadataKind::EnumNoFields5 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(5), false)
-        }
-        IoTypeMetadataKind::EnumNoFields6 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(6), false)
-        }
-        IoTypeMetadataKind::EnumNoFields7 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(7), false)
-        }
-        IoTypeMetadataKind::EnumNoFields8 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(8), false)
-        }
-        IoTypeMetadataKind::EnumNoFields9 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(9), false)
-        }
-        IoTypeMetadataKind::EnumNoFields10 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(10), false)
-        }
-        IoTypeMetadataKind::EnumNoFields11 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(11), false)
-        }
-        IoTypeMetadataKind::EnumNoFields12 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(12), false)
-        }
-        IoTypeMetadataKind::EnumNoFields13 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(13), false)
-        }
-        IoTypeMetadataKind::EnumNoFields14 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(14), false)
-        }
-        IoTypeMetadataKind::EnumNoFields15 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(15), false)
-        }
-        IoTypeMetadataKind::EnumNoFields16 => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-            compact_enum(input, output, Some(16), false)
-        }
-        IoTypeMetadataKind::Array8b => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1 + 1));
-            compact_metadata(input, output)
-        }
-        IoTypeMetadataKind::Array16b => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1 + 2));
-            compact_metadata(input, output)
-        }
-        IoTypeMetadataKind::Array32b => {
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1 + 4));
-            compact_metadata(input, output)
-        }
-        IoTypeMetadataKind::ArrayU8x8
-        | IoTypeMetadataKind::ArrayU8x16
-        | IoTypeMetadataKind::ArrayU8x32
-        | IoTypeMetadataKind::ArrayU8x64
-        | IoTypeMetadataKind::ArrayU8x128
-        | IoTypeMetadataKind::ArrayU8x256
-        | IoTypeMetadataKind::ArrayU8x512
-        | IoTypeMetadataKind::ArrayU8x1024
-        | IoTypeMetadataKind::ArrayU8x2028
-        | IoTypeMetadataKind::ArrayU8x4096 => copy_n_bytes(input, output, 1),
-        IoTypeMetadataKind::VariableBytes8b => copy_n_bytes(input, output, 1 + 1),
-        IoTypeMetadataKind::VariableBytes16b => copy_n_bytes(input, output, 1 + 2),
-        IoTypeMetadataKind::VariableBytes32b => copy_n_bytes(input, output, 1 + 4),
-        IoTypeMetadataKind::VariableBytes512
-        | IoTypeMetadataKind::VariableBytes1024
-        | IoTypeMetadataKind::VariableBytes2028
-        | IoTypeMetadataKind::VariableBytes4096
-        | IoTypeMetadataKind::VariableBytes8192
-        | IoTypeMetadataKind::VariableBytes16384
-        | IoTypeMetadataKind::VariableBytes32768
-        | IoTypeMetadataKind::VariableBytes65536
-        | IoTypeMetadataKind::VariableBytes131072
-        | IoTypeMetadataKind::VariableBytes262144
-        | IoTypeMetadataKind::VariableBytes524288
-        | IoTypeMetadataKind::VariableBytes1048576
-        | IoTypeMetadataKind::VariableBytes2097152
-        | IoTypeMetadataKind::VariableBytes4194304
-        | IoTypeMetadataKind::VariableBytes8388608
-        | IoTypeMetadataKind::VariableBytes16777216 => copy_n_bytes(input, output, 1),
-    }
-}
-
-const fn compact_struct<'i, 'o>(
-    mut input: &'i [u8],
-    mut output: &'o mut [u8],
-    arguments_count: Option<u8>,
-    tuple: bool,
-) -> Option<(&'i [u8], &'o mut [u8])> {
-    if input.is_empty() || output.is_empty() {
-        return None;
-    }
-
-    // Remove struct name
-    let struct_name_length = input[0] as usize;
-    output[0] = 0;
-    (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-    input = forward_option!(skip_n_bytes(input, struct_name_length));
-
-    let mut arguments_count = match arguments_count {
-        Some(arguments_count) => arguments_count,
-        None => {
-            if input.is_empty() || output.is_empty() {
-                return None;
-            }
-
-            let arguments_count = input[0];
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-
-            arguments_count
-        }
-    };
-
-    // Compact arguments
-    while arguments_count > 0 {
-        if input.is_empty() || output.is_empty() {
-            return None;
-        }
-
-        // Remove field name if needed
-        if !tuple {
-            let field_name_length = input[0] as usize;
-            input = forward_option!(skip_n_bytes(input, 1 + field_name_length));
-        }
-
-        // Compact argument's type
-        (input, output) = forward_option!(compact_metadata(input, output));
-
-        arguments_count -= 1;
-    }
-
-    Some((input, output))
-}
-
-const fn compact_enum<'i, 'o>(
-    mut input: &'i [u8],
-    mut output: &'o mut [u8],
-    variant_count: Option<u8>,
-    has_fields: bool,
-) -> Option<(&'i [u8], &'o mut [u8])> {
-    if input.is_empty() || output.is_empty() {
-        return None;
-    }
-
-    // Remove enum name
-    let enum_name_length = input[0] as usize;
-    output[0] = 0;
-    (input, output) = forward_option!(skip_n_bytes_io(input, output, 1));
-    input = forward_option!(skip_n_bytes(input, enum_name_length));
-
-    let mut variant_count = match variant_count {
-        Some(variant_count) => variant_count,
-        None => {
-            if input.is_empty() || output.is_empty() {
-                return None;
-            }
-
-            let variant_count = input[0];
-            (input, output) = forward_option!(copy_n_bytes(input, output, 1));
-
-            variant_count
-        }
-    };
-
-    // Compact enum variants
-    while variant_count > 0 {
-        if input.is_empty() || output.is_empty() {
-            return None;
-        }
-
-        // Remove variant name
-        let field_name_length = input[0] as usize;
-        input = forward_option!(skip_n_bytes(input, 1 + field_name_length));
-
-        if has_fields {
-            // Compact variant as if it was a struct
-            (input, output) = forward_option!(compact_struct(input, output, None, false));
-        } else {
-            // Compact variant as if it was a struct without fields
-            (input, output) = forward_option!(compact_struct(input, output, Some(0), false));
-        }
-
-        variant_count -= 1;
-    }
-
-    Some((input, output))
-}
-
-/// Copies `n` bytes from input to output and returns both input and output after `n` bytes offset
-const fn copy_n_bytes<'i, 'o>(
-    mut input: &'i [u8],
-    mut output: &'o mut [u8],
-    n: usize,
-) -> Option<(&'i [u8], &'o mut [u8])> {
-    if n > input.len() || n > output.len() {
-        return None;
-    }
-
-    let source;
-    let target;
-    (source, input) = input.split_at(n);
-    (target, output) = output.split_at_mut(n);
-    // TODO: Switch to `copy_from_slice` once stable:
-    //  https://github.com/rust-lang/rust/issues/131415
-    // The same as `target.copy_from_slice(&source);`, but it doesn't work in const environment
-    // yet
-    // SAFETY: Size is correct due to slicing above, pointers are created from valid independent
-    // slices of equal length
-    unsafe {
-        ptr::copy_nonoverlapping(source.as_ptr(), target.as_mut_ptr(), source.len());
-    }
-
-    Some((input, output))
-}
-
-/// Skips `n` bytes and return remainder
-const fn skip_n_bytes(input: &[u8], n: usize) -> Option<&[u8]> {
-    if n > input.len() {
-        return None;
-    }
-
-    // `&input[n..]` not supported in const yet
-    Some(input.split_at(n).1)
-}
-
-/// Skips `n` bytes in input and output
-const fn skip_n_bytes_io<'i, 'o>(
-    mut input: &'i [u8],
-    mut output: &'o mut [u8],
-    n: usize,
-) -> Option<(&'i [u8], &'o mut [u8])> {
-    if n > input.len() || n > output.len() {
-        return None;
-    }
-
-    // `&input[n..]` not supported in const yet
-    input = input.split_at(n).1;
-    // `&mut output[n..]` not supported in const yet
-    output = output.split_at_mut(n).1;
-
-    Some((input, output))
 }
